Default approval status

+++ machina/apps/forum_conversation/abstract_models.py
@@ -204,9 +204,9 @@ class AbstractTopic(models.Model):
 
     def update_trackers(self):
         """ Updates the denormalized trackers associated with the topic instance. """
-        self.posts_count = self.posts.filter(approved=True).count()
+        self.posts_count = self.posts.exclude(approved=False).count()
         first_post = self.posts.all().order_by('created').first()
-        last_post = self.posts.filter(approved=True).order_by('-created').first()
+        last_post = self.posts.exclude(approved=False).order_by('-created').first()
         self.first_post = first_post
         self.last_post = last_post
         self.last_post_on = last_post.created if last_post else None
@@ -246,7 +246,7 @@ class AbstractPost(DatedModel):
     username = models.CharField(max_length=155, blank=True, null=True, verbose_name=_('Username'))
 
     # A post can be approved before publishing ; defaults to True
-    approved = models.BooleanField(default=True, db_index=True, verbose_name=_('Approved'))
+    approved = models.BooleanField(default=machina_settings.DEFAULT_APPROVAL_STATUS, db_index=True, verbose_name=_('Approved'))
 
     # The user can choose if they want to display their signature with the content of the post
     enable_signature = models.BooleanField(
@@ -338,10 +338,15 @@ class AbstractPost(DatedModel):
 
     def delete(self, using=None):
         """ Deletes the post instance. """
-        if self.is_alone:
-            # The default way of operating is to trigger the deletion of the associated topic
-            # only if the considered post is the only post embedded in the topic
-            self.topic.delete()
-        else:
-            super(AbstractPost, self).delete(using)
+        if machina_settings.DEFAULT_APPROVAL_STATUS is None and self.approved is not False:
+            self.approved = False
+            self.save(update_fields=['approved'])
             self.topic.update_trackers()
+        else:
+            if self.is_alone:
+                # The default way of operating is to trigger the deletion of the associated topic
+                # only if the considered post is the only post embedded in the topic
+                self.topic.delete()
+            else:
+                super(AbstractPost, self).delete(using)
+                self.topic.update_trackers()
